requirements.txt
beautifulsoup4==4.4.1
bleach==1.4.2
chardet==2.3.0
Django==1.9.6
django-contrib-comments==1.7.1
django-object-tools==1.9
django-sortedm2m==1.2.2
easy-thumbnails==2.3
filebrowser-safe==0.4.3
future==0.15.2
grappelli-safe==0.4.2
html5lib==0.9999999
Mezzanine==4.1.0
oauthlib==1.1.1
Pillow==3.2.0
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
tzlocal==1.2.2
wheel==0.24.0
